requirements.txt
-c constraints.txt

tornado==6.0.2 \
    --hash=sha256:1174dcb84d08887b55defb2cda1986faeeea715fff189ef3dc44cce99f5fca6b \
    --hash=sha256:2613fab506bd2aedb3722c8c64c17f8f74f4070afed6eea17f20b2115e445aec \
    --hash=sha256:44b82bc1146a24e5b9853d04c142576b4e8fa7a92f2e30bc364a85d1f75c4de2 \
    --hash=sha256:457fcbee4df737d2defc181b9073758d73f54a6cfc1f280533ff48831b39f4a8 \
    --hash=sha256:49603e1a6e24104961497ad0c07c799aec1caac7400a6762b687e74c8206677d \
    --hash=sha256:8c2f40b99a8153893793559919a355d7b74649a11e59f411b0b0a1793e160bc0 \
    --hash=sha256:e1d897889c3b5a829426b7d52828fb37b28bc181cd598624e65c8be40ee3f7fa
wsaccel==0.6.2 \
    --hash=sha256:425706acf0724d2f6bfa391ec37b4ef121d3432c956029de3cea4e101c218e0c
ujson==1.35 \
    --hash=sha256:f66073e5506e91d204ab0c614a148d5aa938bdbf104751be66f8ad7a222f5f86
Django==2.2.1 \
    --hash=sha256:6fcc3cbd55b16f9a01f37de8bcbe286e0ea22e87096557f1511051780338eaea \
    --hash=sha256:bb407d0bb46395ca1241f829f5bd03f7e482f97f7d1936e26e98dacb201ed4ec
dj-database-url==0.5.0 \
    --hash=sha256:4aeaeb1f573c74835b0686a2b46b85990571159ffc21aa57ecd4d1e1cb334163 \
    --hash=sha256:851785365761ebe4994a921b433062309eb882fedd318e1b0fcecc607ed02da9
python-decouple==3.1 \
    --hash=sha256:1317df14b43efee4337a4aa02914bf004f010cd56d6c4bd894e6474ec8c4fe2d
psycopg2-binary==2.8.2 \
    --hash=sha256:007ca0df127b1862fc010125bc4100b7a630efc6841047bd11afceadb4754611 \
    --hash=sha256:03c49e02adf0b4d68f422fdbd98f7a7c547beb27e99a75ed02298f85cb48406a \
    --hash=sha256:0a1232cdd314e08848825edda06600455ad2a7adaa463ebfb12ece2d09f3370e \
    --hash=sha256:131c80d0958c89273d9720b9adf9df1d7600bb3120e16019a7389ab15b079af5 \
    --hash=sha256:2de34cc3b775724623f86617d2601308083176a495f5b2efc2bbb0da154f483a \
    --hash=sha256:2eddc31500f73544a2a54123d4c4b249c3c711d31e64deddb0890982ea37397a \
    --hash=sha256:484f6c62bdc166ee0e5be3aa831120423bf399786d1f3b0304526c86180fbc0b \
    --hash=sha256:4c2d9369ed40b4a44a8ccd6bc3a7db6272b8314812d2d1091f95c4c836d92e06 \
    --hash=sha256:70f570b5fa44413b9f30dbc053d17ef3ce6a4100147a10822f8662e58d473656 \
    --hash=sha256:7a2b5b095f3bd733aab101c89c0e1a3f0dfb4ebdc26f6374805c086ffe29d5b2 \
    --hash=sha256:804914a669186e2843c1f7fbe12b55aad1b36d40a28274abe6027deffad9433d \
    --hash=sha256:8520c03172da18345d012949a53617a963e0191ccb3c666f23276d5326af27b5 \
    --hash=sha256:90da901fc33ea393fc644607e4a3916b509387e9339ec6ebc7bfded45b7a0ae9 \
    --hash=sha256:a582416ad123291a82c300d1d872bdc4136d69ad0b41d57dc5ca3df7ef8e3088 \
    --hash=sha256:ac8c5e20309f4989c296d62cac20ee456b69c41fd1bc03829e27de23b6fa9dd0 \
    --hash=sha256:b2cf82f55a619879f8557fdaae5cec7a294fac815e0087c4f67026fdf5259844 \
    --hash=sha256:b59d6f8cfca2983d8fdbe457bf95d2192f7b7efdb2b483bf5fa4e8981b04e8b2 \
    --hash=sha256:be08168197021d669b9964bd87628fa88f910b1be31e7010901070f2540c05fd \
    --hash=sha256:be0f952f1c365061041bad16e27e224e29615d4eb1fb5b7e7760a1d3d12b90b6 \
    --hash=sha256:c1c9a33e46d7c12b9c96cf2d4349d783e3127163fd96254dcd44663cf0a1d438 \
    --hash=sha256:d18c89957ac57dd2a2724ecfe9a759912d776f96ecabba23acb9ecbf5c731035 \
    --hash=sha256:d7e7b0ff21f39433c50397e60bf0995d078802c591ca3b8d99857ea18a7496ee \
    --hash=sha256:da0929b2bf0d1f365345e5eb940d8713c1d516312e010135b14402e2a3d2404d \
    --hash=sha256:de24a4962e361c512d3e528ded6c7480eab24c655b8ca1f0b761d3b3650d2f07 \
    --hash=sha256:e45f93ff3f7dae2202248cf413a87aeb330821bf76998b3cf374eda2fc893dd7 \
    --hash=sha256:f046aeae1f7a845041b8661bb7a52449202b6c5d3fb59eb4724e7ca088811904 \
    --hash=sha256:f1dc2b7b2748084b890f5d05b65a47cd03188824890e9a60818721fd492249fb \
    --hash=sha256:fcbe7cf3a786572b73d2cd5f34ed452a5f5fac47c9c9d1e0642c457a148f9f88
geoip2==2.9.0 \
    --hash=sha256:a37ddac2d200ffb97c736da8b8ba9d5d8dc47da6ec0f162a461b681ecac53a14 \
    --hash=sha256:f7ffe9d258e71a42cf622ce6350d976de1d0312b9f2fbce3975c7d838b57ecf0
requests[security]==2.21.0 \
    --hash=sha256:502a824f31acdacb3a35b6690b5fbf0bc41d63a24a45c4004352b0242707598e \
    --hash=sha256:7bf2a778576d825600030a110f3c0e3e8edc51dfaafe1c146e39a2027784957b
maxminddb==1.4.1 \
    --hash=sha256:df1451bcd848199905ac0de4631b3d02d6a655ad28ba5e5a4ca29a23358db712
